scripts/requirements.txt
supabase
python-dotenv
requests
httpx
beautifulsoup4
